+++ EmployeeLeaveManagementWebAPI/Domain/LeaveReportModel.cs
@@ -10,18 +10,18 @@ namespace LMS_WebAPI_Domain
     {
         public int Year { get; set; }
 
-        public int Jan { get; set; }
-        public int Feb { get; set; }
-        public int Mar { get; set; }
-        public int Apr { get; set; }
-        public int May { get; set; }
-        public int Jun { get; set; }
-        public int Jul { get; set; }
-        public int Aug { get; set; }
-        public int Sep { get; set; }
-        public int Oct { get; set; }
-        public int Nov { get; set; }
-        public int Dec { get; set; }
+        public double Jan { get; set; }
+        public double Feb { get; set; }
+        public double Mar { get; set; }
+        public double Apr { get; set; }
+        public double May { get; set; }
+        public double Jun { get; set; }
+        public double Jul { get; set; }
+        public double Aug { get; set; }
+        public double Sep { get; set; }
+        public double Oct { get; set; }
+        public double Nov { get; set; }
+        public double Dec { get; set; }
         public int LeaveCount { get; set; }
 
     }
